All classes commented with full javadoc-compatible comments and parameters.  package-info.java files added to each package within the main java directory, providing descriptions of the functions of each package's classes.
Javadoc has been generated within the webapp folder, and linked to the navigation module of the page.  Currently set up to display in a new window, and generated to include full resources for as complete a document as possible.  As everything is in the webapp folder, it should generate in the war file for deployment to AWS with all documentation intact.